[TASK] Make TYPO3\FLOW3\var_dump() available in unit tests
FLOW3's var_dump() function could not be used in unit tests
because the class loader has no way to derive the path and
filename from the function name. In a non-testing context
that's no problem as the containing file is loaded anyway
but that was not the case in a unit testing context.

Now the Debugger class is loaded in the unit test bootstrap.

Change-Id: Ibf07759cdbb64459dbc2ef04c0af91d31851f2a1
Releases: 1.1