[TASK] Use custom phantomjs startup script
THe phantomjs startup script in buster.js does not check for a
variable to be set before calling the split() method on this variable.
By creating our own phantomjs startup script we can work around this
error and gain more control on the phantomjs startup.

This change also adds some hints to the documentation.

Related: #41508
Releases: 1.2

Change-Id: Ic40fc1c4a71d0aef225cad04be959d937b14864c