[TASK] Custom context support for functional tests
To be able to run tests with a given context the bootstrap is looking
for a configured context. if none is given the default is used. The
context is restricted to "Testing" and its subcontexts "Testing/*".

Example usage::

 FLOW_CONTEXT=Testing/Postgres bin/phpunit -c Build/BuildEssentials/PhpUnit/FunctionalTests.xml

This is useful to test different behaviors for different types of dbs.

Resolves: FLOW-218
Releases: master, 3.0
Change-Id: I7b32e1fb21b47680199846c76c4e782910a6db14