[TASK] Adjust Crowdin configuration file template
This changes the Crowdin configuration to allow the use of multiple
languages that would map to the same two letter code (like pt-PT and
pt-BR).

It does that by using the locale instead of the two letter language
code and defining a mapping for the existing translations to stay
unchanged.

See https://discuss.neos.io/t/510/ for more details.